hax debug

+++ server.cpp
@@ -122,7 +122,8 @@ std::string getClientData(epoll_event &event, std::vector <Socket> &clients) {
       break;
     }
     uint16_t len = fromNetworkByteOrder(shortFromChars(buffer));
-    debug() << "LEN: " << (int) len << " " << shortFromChars(buffer) << " " << (int) buffer[0] << " " << (int) buffer[1] << "\n";
+    debug() << "LEN: " << (int) len << " " << (int) shortFromChars(buffer) << " " << (int) buffer[0] << " " <<
+    (int) buffer[1] << "\n";
     buffer[0] = buffer[1] = '#';
     buffer[count] = '\0';
     std::string tmp(buffer);
